feat(cli): single `hydrozoa` dispatcher over all entrypoints
Fold the server and every bootstrap/deploy/demo main into one decline
dispatcher (`hydrozoa <subcommand>`); the old per-command CommandIOApps become
`Command` values, the server object is `Serve` (subcommand `serve`), and the
interactive demos move from the retired `examples` project into `hydrozoa.app.cli`.

- BuildInfo: bake version + git revision + build time into `hydrozoa.BuildInfo`;
  logged at `serve` startup, printed by `hydrozoa version` / `--version`, and
  served from the new `GET /version`.
- Clean output: mute Bloxbean's per-request DEBUG flood and pass the
  native-access JVM flags to `run`/the packaged launcher (no sbt prefixes, no
  Blockfrost noise, no restricted-method warnings).
- Packaging: name the launcher `hydrozoa`, emit only the dispatcher script,
  JDK-25 Docker base, and stamp the image with the git revision label.
