+++ build.sbt
@@ -1,16 +1,36 @@
 enablePlugins(
   JavaAppPackaging,
-  DockerPlugin
+  DockerPlugin,
+  BuildInfoPlugin
 )
 
 Compile / mainClass := Some("hydrozoa.app.Main")
+// Name the packaged launcher `hydrozoa`, and generate only the dispatcher's script (not forwarder
+// scripts for every other discovered main); every command is a `hydrozoa <subcommand>`.
+executableScriptName := "hydrozoa"
+Compile / discoveredMainClasses := Seq("hydrozoa.app.Main")
+
+// The git revision baked into the Docker image labels; matches `hydrozoa.BuildInfo.gitCommit`.
+lazy val gitRevision: String =
+    scala.util
+        .Try(scala.sys.process.Process("git describe --always --dirty --abbrev=8").!!.trim)
+        .getOrElse("unknown")
+
+// Bake the clean-output JVM flags into the generated launcher scripts (both `stage` and the Docker
+// image), matching the `run` / `Test` settings: silence blst-java's JNI `System::load` and Scala
+// `LazyVals`' `sun.misc.Unsafe` restricted-method warnings.
+bashScriptExtraDefines ++= Seq(
+  """addJava "--enable-native-access=ALL-UNNAMED"""",
+  """addJava "--sun-misc-unsafe-memory-access=allow""""
+)
 
 // Docker settings
 Docker / packageName := "cardano-hydrozoa/hydrozoa"
 Docker / version := version.value
 Docker / daemonUser := "hydrozoa"
 Docker / daemonGroup := "hydrozoa"
-dockerBaseImage := "eclipse-temurin:21-jre-jammy" // Use Debian-based image for better compatibility
+// JDK 25 to match the project's language/runtime flags (--sun-misc-unsafe-memory-access is 23+).
+dockerBaseImage := "eclipse-temurin:25-jre"
 dockerExposedPorts ++= Seq(8080)
 
 // Skip documentation generation for Docker
@@ -23,7 +43,8 @@ Global / excludeLintKeys += Docker / dockerEnvVars
 Docker / dockerLabels := Map(
   "org.opencontainers.image.title" -> "Hydrozoa",
   "org.opencontainers.image.description" -> "Cardano Hydrozoa L2 State Channel",
-  "org.opencontainers.image.version" -> version.value
+  "org.opencontainers.image.version" -> version.value,
+  "org.opencontainers.image.revision" -> gitRevision
 )
 
 Docker / dockerEnvVars := Map(
@@ -138,8 +159,30 @@ lazy val core: Project = (project in file("."))
         "dev.optics" %% "monocle-macro" % "3.3.0" % Test,
         "co.fs2" %% "fs2-io" % "3.12.2" % Test
       ),
+      // Bake the version, git revision, and build time into `hydrozoa.BuildInfo` so they can be
+      // logged at startup, served from `GET /version`, and stamped onto the Docker image labels.
+      buildInfoPackage := "hydrozoa",
+      buildInfoKeys := Seq[BuildInfoKey](
+        version,
+        BuildInfoKey.action("gitCommit") {
+            scala.util
+                .Try(scala.sys.process.Process("git describe --always --dirty --abbrev=8").!!.trim)
+                .getOrElse("unknown")
+        }
+      ),
+      // BuildTime forces a regenerate each build, keeping gitCommit current within a warm session.
+      buildInfoOptions += BuildInfoOption.BuildTime,
       // Fork JVM to properly pass system properties
       run / fork := true,
+      // Silence the JVM's restricted-method warnings (blst-java JNI `System::load`, Scala
+      // `LazyVals` via `sun.misc.Unsafe`) for forked `run`/CLI invocations, matching Test.
+      run / javaOptions ++= Seq(
+        "--enable-native-access=ALL-UNNAMED",
+        "--sun-misc-unsafe-memory-access=allow"
+      ),
+      // The interactive `submit-deposit` / `submit-l2-tx` subcommands read console prompts; wire
+      // stdin through to the forked `sbt run` JVM. The packaged launcher gets stdin natively.
+      run / connectInput := true,
       // Fork each test run into a fresh JVM: isolates native state (RocksDB JNI), the
       // cats-effect IORuntime, and daemon threads, and makes re-running a `testOnly` in a
       // warm sbt session actually re-run instead of reporting 0 tests.
@@ -154,20 +197,6 @@ lazy val core: Project = (project in file("."))
         "--enable-native-access=ALL-UNNAMED",
         "--sun-misc-unsafe-memory-access=allow"
       ),
-    )
-
-// Interactive demo targets that drive a running head (DEPLOYMENT.md §7).
-lazy val examples: Project = (project in file("examples"))
-    .dependsOn(core)
-    .settings(
-      name := "hydrozoa-examples",
-      publish / skip := true,
-      run / fork := true,
-      // Wire stdin through to the forked JVM — the demo targets read prompts from the console.
-      run / connectInput := true,
-      // Forked runs resolve relative paths (config/demo) from the repo root, not examples/.
-      Compile / run / forkOptions := (Compile / run / forkOptions).value
-          .withWorkingDirectory((ThisBuild / baseDirectory).value),
     )
 
 // Integration tests
@@ -216,10 +245,10 @@ addCompilerPlugin("org.scalus" % "scalus-plugin" % scalusVersion cross CrossVers
 //addCommandAlias("fmtCheckAll", ";core/scalafmtCheckAll ;integration/scalafmtCheckAll ;benchmark/scalafmtCheckAll")
 //addCommandAlias("lintAll", ";core/scalafixAll ;integration/scalafixAll ;benchmark/scalafixAll")
 //addCommandAlias("lintCheckAll", ";core/scalafixAll --check ;integration/scalafixAll --check ;benchmark/scalafixAll --check")
-addCommandAlias("fmtAll", ";core/scalafmtAll ;examples/scalafmtAll")
-addCommandAlias("fmtCheckAll", ";core/scalafmtCheckAll ;examples/scalafmtCheckAll")
-addCommandAlias("lintAll", ";core/scalafixAll ;examples/scalafixAll")
-addCommandAlias("lintCheckAll", ";core/scalafixAll --check ;examples/scalafixAll --check")
+addCommandAlias("fmtAll", ";core/scalafmtAll")
+addCommandAlias("fmtCheckAll", ";core/scalafmtCheckAll")
+addCommandAlias("lintAll", ";core/scalafixAll")
+addCommandAlias("lintCheckAll", ";core/scalafixAll --check")
 
 // Test dependencies
 ThisBuild / testFrameworks += new TestFramework("org.scalatest.tools.Framework")
