+++ src/test/resources/logback-test.xml
@@ -10,6 +10,7 @@
     </root>
     <logger name="BlockWeaver" level="warn"/>
     <logger name="CardanoBackend" level="warn"/>
+    <logger name="com.bloxbean" level="warn"/>
     <logger name="CardanoLiaison" level="warn"/>
     <logger name="CardanoLiaisonTest" level="warn"/>
     <logger name="PeerLiaison" level="warn"/>
